refactor: organize directories again
