+++ src/main/main.ts
@@ -11,10 +11,10 @@ import {
 } from './input'
 import { Client as GitHubClient } from './github'
 import { makeMarkdownReport } from './markdown'
-import { JUnitReporterFactory } from '../junit/factory'
-import { TableSetFactory } from '../composer/factory'
+import { JUnitReporterFactory } from '../reporter/factory'
+import { TableSetFactory } from '../table/factory'
 import { ReporterType } from '../type'
-import { JUnitXmlReader } from '../junit/reader'
+import { JUnitXmlReader } from '../reporter/reader'
 
 const mark = '<!-- commented by junit-monorepo-go -->'
