+++ src/reporter/factory.ts
