+++ src/reporter/golangcilint.ts
