+++ src/table/base/typed.ts
