+++ src/table/result.ts
